+++ lib/gst/plugin/handlers/meson.build
@@ -1,0 +1,1 @@
+subdir('gl')
