+++ lib/gst/plugin/importers/gstclappergluploader.h
@@ -19,18 +19,27 @@
 
 #pragma once
 
-#include "gstclapperglbaseimporter.h"
+#include "gst/plugin/gstclapperimporter.h"
+#include "gst/plugin/handlers/gl/gstclapperglcontexthandler.h"
 
 G_BEGIN_DECLS
 
 #define GST_TYPE_CLAPPER_GL_UPLOADER (gst_clapper_gl_uploader_get_type())
-G_DECLARE_FINAL_TYPE (GstClapperGLUploader, gst_clapper_gl_uploader, GST, CLAPPER_GL_UPLOADER, GstClapperGLBaseImporter)
+G_DECLARE_FINAL_TYPE (GstClapperGLUploader, gst_clapper_gl_uploader, GST, CLAPPER_GL_UPLOADER, GstClapperImporter)
 
 #define GST_CLAPPER_GL_UPLOADER_CAST(obj)        ((GstClapperGLUploader *)(obj))
 
+#define GST_CLAPPER_GL_UPLOADER_GET_LOCK(obj)    (&GST_CLAPPER_GL_UPLOADER_CAST(obj)->lock)
+#define GST_CLAPPER_GL_UPLOADER_LOCK(obj)        g_mutex_lock (GST_CLAPPER_GL_UPLOADER_GET_LOCK(obj))
+#define GST_CLAPPER_GL_UPLOADER_UNLOCK(obj)      g_mutex_unlock (GST_CLAPPER_GL_UPLOADER_GET_LOCK(obj))
+
 struct _GstClapperGLUploader
 {
-  GstClapperGLBaseImporter parent;
+  GstClapperImporter parent;
+
+  GMutex lock;
+
+  GstClapperGLContextHandler *gl_handler;
 
   GstGLUpload *upload;
   GstGLColorConvert *color_convert;
